+++ src/main/java/roomescape/reservations/application/ReservationService.java
@@ -1,14 +1,14 @@
-package roomescape.domain.reservations.application;
+package roomescape.reservations.application;
 
 import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import roomescape.domain.reservations.entity.Reservation;
-import roomescape.domain.reservations.entity.ReservationTime;
-import roomescape.domain.reservations.entity.ReservationRepository;
-import roomescape.domain.reservations.entity.ReservationTimeRepository;
-import roomescape.domain.reservations.presentation.dto.ReservationRequest;
-import roomescape.domain.reservations.presentation.dto.ReservationResponse;
+import roomescape.reservations.entity.Reservation;
+import roomescape.reservations.entity.ReservationTime;
+import roomescape.reservations.entity.ReservationRepository;
+import roomescape.reservations.entity.ReservationTimeRepository;
+import roomescape.reservations.presentation.dto.ReservationRequest;
+import roomescape.reservations.presentation.dto.ReservationResponse;
 
 @Service
 public class ReservationService {
