+++ src/main/java/roomescape/reservations/entity/ReservationTimeRepository.java
@@ -1,4 +1,4 @@
-package roomescape.domain.reservations.entity;
+package roomescape.reservations.entity;
 
 import java.util.List;
 import java.util.Optional;
