+++ src/main/java/roomescape/reservations/infrastructure/ReservationTimeJdbcTemplateRepository.java
@@ -1,4 +1,4 @@
-package roomescape.domain.reservations.infrastructure;
+package roomescape.reservations.infrastructure;
 
 import java.util.List;
 import java.util.Map;
@@ -7,8 +7,8 @@ import javax.sql.DataSource;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
-import roomescape.domain.reservations.entity.ReservationTime;
-import roomescape.domain.reservations.entity.ReservationTimeRepository;
+import roomescape.reservations.entity.ReservationTime;
+import roomescape.reservations.entity.ReservationTimeRepository;
 
 @Repository
 public class ReservationTimeJdbcTemplateRepository implements ReservationTimeRepository {
