+++ src/main/java/roomescape/reservations/presentation/ReservationTimeController.java
@@ -1,4 +1,4 @@
-package roomescape.domain.reservations.presentation;
+package roomescape.reservations.presentation;
 
 import java.net.URI;
 import java.util.List;
@@ -9,9 +9,9 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.domain.reservations.application.ReservationTimeService;
-import roomescape.domain.reservations.presentation.dto.ReservationTimeRequest;
-import roomescape.domain.reservations.presentation.dto.ReservationTimeResponse;
+import roomescape.reservations.application.ReservationTimeService;
+import roomescape.reservations.presentation.dto.ReservationTimeRequest;
+import roomescape.reservations.presentation.dto.ReservationTimeResponse;
 
 @RestController
 public class ReservationTimeController {
